Split tickets list context
