+++ domains/eventEditor/src/services/context/EntityListContext/tickets/TicketsListProvider.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { TicketsFilterStateProvider } from './TicketsFilterStateProvider';
+import { FilteredTicketsProvider } from './FilteredTicketsProvider';
+
+export const TicketsListProvider: React.FC = ({ children }) => {
+	return (
+		<TicketsFilterStateProvider>
+			<FilteredTicketsProvider>{children}</FilteredTicketsProvider>
+		</TicketsFilterStateProvider>
+	);
+};
